+++ src/main/java/rtg/world/biome/realistic/RealisticBiomeBase.java
@@ -11,6 +11,7 @@ import rtg.api.RTGAPI;
 import rtg.api.config.BiomeConfig;
 import rtg.api.config.RTGConfig;
 import rtg.api.util.Accessor;
+import rtg.api.util.BiomeUtils;
 import rtg.api.util.Logger;
 import rtg.api.util.noise.OpenSimplexNoise;
 import rtg.api.util.noise.SimplexCellularNoise;
@@ -26,7 +27,6 @@ import rtg.api.world.surface.SurfaceGeneric;
 import rtg.api.world.surface.SurfaceRiverOasis;
 import rtg.api.world.terrain.TerrainBase;
 import rtg.api.world.terrain.TerrainOrganic;
-import rtg.world.biome.BiomeAnalyzer;
 
 
 @SuppressWarnings({"WeakerAccess", "UnusedParameters", "unused"})
@@ -164,7 +164,7 @@ public abstract class RealisticBiomeBase implements IRealisticBiome {
      * Returns the beach biome to use for this biome, with a dynamically-calculated preferred beach.
      */
     public Biome beachBiome() {
-        return this.beachBiome(BiomeAnalyzer.getPreferredBeachForBiome(this.baseBiome));
+        return this.beachBiome(BiomeUtils.getPreferredBeachForBiome(this.baseBiome));
     }
 
     public BiomeDecoratorRTG rDecorator() {
